+++ example/app/example/index.tsx
@@ -23,6 +23,7 @@ const examples: Example[] = [
       { name: "Form", route: "/example/composable-screen-form" },
       { name: "Carousel", route: "/example/composable-screen-carousel" },
       { name: "Button", route: "/example/composable-screen-button" },
+      { name: "ZStack", route: "/example/composable-screen-zstack" },
       { name: "All", route: "/example/composable-screen" },
     ],
   },